[knife] Use password prompt for use-sudo-password
If the user has requested the use of sudo, and use-sudo-password, but they have
not provided the password on the command line, knife will prompt for the
password to connect via SSH but not use this password in the sudo command. This
change prompts the user for the password early enough in the command processing
to be present for use with sudo. The password collected will also be used for
the SSH connection.